wip: update import attribute types
