+++ v3/src/components/data-display/components/legend/legend-attribute-label.tsx
@@ -1,6 +1,6 @@
 import React, {useCallback, useEffect, useRef} from "react"
 import {select} from "d3"
-import {AttributeType} from "../../../../models/data/attribute"
+import {AttributeType} from "../../../../models/data/attribute-types"
 import {IDataSet} from "../../../../models/data/data-set"
 import {axisGap} from "../../../axis/axis-types"
 import {GraphPlace} from "../../../axis-graph-shared"
