+++ v3/src/components/data-display/models/data-configuration-model.ts
@@ -8,7 +8,7 @@ import {applyModelChange} from "../../../models/history/apply-model-change"
 import {cachedFnWithArgsFactory} from "../../../utilities/mst-utils"
 import { isFiniteNumber } from "../../../utilities/math-utils"
 import { stringValuesToDateSeconds } from "../../../utilities/date-utils"
-import {AttributeType, attributeTypes} from "../../../models/data/attribute"
+import {AttributeType, attributeTypes} from "../../../models/data/attribute-types"
 import {DataSet, IDataSet} from "../../../models/data/data-set"
 import {ICase} from "../../../models/data/data-set-types"
 import {idOfChildmostCollectionForAttributes} from "../../../models/data/data-set-utils"
