+++ v3/src/components/graph/components/graph-attribute-label.tsx
@@ -4,7 +4,7 @@ import { t } from "../../../utilities/translation/translate"
 import { useGraphDataConfigurationContext } from "../hooks/use-graph-data-configuration-context"
 import { useGraphContentModelContext } from "../hooks/use-graph-content-model-context"
 import { useGraphLayoutContext } from "../hooks/use-graph-layout-context"
-import { AttributeType } from "../../../models/data/attribute"
+import { AttributeType } from "../../../models/data/attribute-types"
 import { IDataSet } from "../../../models/data/data-set"
 import { GraphPlace, isVertical } from "../../axis-graph-shared"
 import { AttributeLabel } from "../../data-display/components/attribute-label"
