+++ v3/src/components/graph/components/graph-axis.tsx
@@ -10,7 +10,7 @@ import {useInstanceIdContext} from "../../../hooks/use-instance-id-context"
 import {useGraphDataConfigurationContext} from "../hooks/use-graph-data-configuration-context"
 import {useGraphContentModelContext} from "../hooks/use-graph-content-model-context"
 import {useGraphLayoutContext} from "../hooks/use-graph-layout-context"
-import {AttributeType} from "../../../models/data/attribute"
+import {AttributeType} from "../../../models/data/attribute-types"
 import {IDataSet} from "../../../models/data/data-set"
 import {AxisPlace} from "../../axis/axis-types"
 import {Axis} from "../../axis/components/axis"
