+++ v3/src/models/data/attribute.ts
@@ -28,7 +28,6 @@
 import { Instance, SnapshotIn, types } from "mobx-state-tree"
 import { kAttrIdPrefix, typeV3Id } from "../../utilities/codap-utils"
 import { parseColor } from "../../utilities/color-utils"
-import { formatStdISODateString } from "../../utilities/date-iso-utils"
 import { isDateString } from "../../utilities/date-parser"
 import { DatePrecision } from "../../utilities/date-utils"
 import { cachedFnFactory } from "../../utilities/mst-utils"
@@ -36,32 +35,13 @@ import { isBoundaryValue, kPolygonNames } from "../boundaries/boundary-types"
 import { Formula, IFormula } from "../formula/formula"
 import { applyModelChange } from "../history/apply-model-change"
 import { withoutUndo } from "../history/without-undo"
-import { isDevelopment, isProduction, IValueType } from "./attribute-types"
+import {
+  AttributeType, attributeTypes, importValueToString, isDevelopment, isProduction, IValueType
+} from "./attribute-types"
 import { V2Model } from "./v2-model"
 
 export interface ISetValueOptions {
   noInvalidate?: boolean
-}
-
-export function importValueToString(value: IValueType): string {
-  if (value == null) {
-    return ""
-  }
-  if (typeof value === "string") {
-    return value
-  }
-  if (value instanceof Date) {
-    return formatStdISODateString(value)
-  }
-  return value.toString()
-}
-
-export const attributeTypes = [
-  "categorical", "numeric", "date", "qualitative", "boundary", "checkbox", "color"
-] as const
-export type AttributeType = typeof attributeTypes[number]
-export function isAttributeType(type?: string | null): type is AttributeType {
-  return type != null && (attributeTypes as readonly string[]).includes(type)
 }
 
 export const Attribute = V2Model.named("Attribute").props({
